Refactor to use `Requests<http://docs.python-requests.org/`_
- Create ``setup.py`` in so that the library is installable via
  easy_install, pip, python setup.py install, etc.

- Move ``urllib2_file.py`` into ``youtrack`` package so that it is
  included when installed

- Fix erroneous imports in ``connection_test.py``

With regard to the ``connection_test.py``, the tests run, but fail or
error.  Most of these are due to invalid login and or missing data in
the test instance.  More unit tests need to be added, as well as a
default "test" instance provided so that the functional tests will work.